Add assembly version of simple operations on aarch64
For aarch64 and arm64ec with Neon, add assembly versions of the
following:

* `ceil`
* `ceilf`
* `fabs`
* `fabsf`
* `floor`
* `floorf`
* `fma`
* `fmaf`
* `round`
* `roundf`
* `sqrt`
* `sqrtf`
* `trunc`
* `truncf`

If the `fp16` target feature is available, which implies `neon`, also
include the following:

* `ceilf16`
* `fabsf16`
* `floorf16`
* `rintf16`
* `roundf16`
* `sqrtf16`
* `truncf16`

Additionally, replace `core::arch` versions of the following with
handwritten assembly (which avoids issues with `aarch64be`):

* `rint`
* `rintf`

Instructions for `fmax` and `fmin` are also available but seem to
provide different results based on whether NaN inputs are signaling or
quiet. Our current implementation does not do this, so omit these for
now.